hotfix: changed filter heading copy & renamed components from 🇳🇱 to 🏴󠁧󠁢󠁥󠁮󠁧󠁿
